Add split-block Bloom filters.
This adds split-block Bloom filters, which are identical to
traditional Bloom filters in their interface, but which limit memory
reads and writes to a small (16-64 byte) contiguous region of memory.

Unlike traditional Bloom filters, this implementation here is limited
to 4, 8, or 16 bits that are accessed on each operation. Additionally,
those bits are all within lanes, i.e. no two accessed bits ever
overlap. These two features allow the use of SIMD instructions, which
speeds up the implementation considerably when the filter is in fast
memory.

This algorithm is from Apache Impala, and has also been adopted by
Apache Kudu and Apache Parquet.