Build local files with static library
Fo system installation, the shared object might be more useful, but
after linking against it, running any executables in this project
required knowledge of where to find it - and not to find the one
installed in the system folder!

Instead, we'll build both static and dynamic and link against the
static one for local projects.