128-bit lane-wise-shift is not in any SSE ISA
As part of the mask creation, we need to shift a one in each lane left
by the value of the corresponding lane in another 128-bit
register. This is actually not available in SSE. I was using
_mm_sll_epi32, but that doesn't treat the second argument as lanes,
but as an extended uint64_t.

This commit fixes that, but the ISA requirement for 4x32 has gone up
to AVX2.