Add Simd-Scalar equivalence test
The Simd and Scalar filter implementations should be identical to
ensure that there are no semantic differences between filters that
were run on different machines. SOmething similar to this actually
came up in Apache Impala; see
https://issues.apache.org/jira/browse/IMPALA-3817.

This patch also adds comments to existing tests and also adds
randomization to them.