Change Makefiles to detect needed headers
This is change that makes the Mkaefiles longer, but also hopefully
makes them more likely to be correct. This uses GCC's (and Clang's -MM
flag) to detect local headers a file depends on.

I do not believe this takes care of all transitive dependencies,
unfortunately, and it requires different rules for compiling C and C++
files.